init commit: add manifest.json, github.js and swagger.js
github.js: for .yaml files in GitHub adds "Open in Swagger" button with
Swagger Editor link. In addition, it stores the file content to
LocalStorage with chrome.storage API.

swagger.js: if there is a value by "openapi" key in the LocalStorage,
then it creates a file from it, sets it to the file input element and
triggers "change" event to apply it.

As a result, when the user clicks "Open in Swagger", the contents on the
file is opened in Swagger Editor.